refactor: simplify the components
